+++ packages/widget/src/components/common/dropdown/dropdown.ts
@@ -135,7 +135,7 @@ export class Dropdown extends BaseComponent {
       this.selectedOption,
       () =>
         html`${this.selectedOption!.icon || networkIconsMap.default}
-          <span part="optionName" class="optionName">
+          <span part="optionName" class="optionName optionNameEllipsis">
             ${capitalize(this.selectedOption!.name)}
           </span>`,
       () =>
@@ -156,7 +156,7 @@ export class Dropdown extends BaseComponent {
       () => customOptionHtml,
       () => html`
         ${icon || ''}
-        <span class="optionName">${capitalize(name)}</span>
+        <span class="optionName optionNameEllipsis">${capitalize(name)}</span>
       `
     );
   }