+++ packages/widget/src/components/transfer/fungible/fungible-token-transfer.ts
@@ -1,11 +1,10 @@
 import type { Domain } from '@buildwithsygma/sygma-sdk-core';
-import { Environment } from '@buildwithsygma/sygma-sdk-core';
+import { Environment, Network } from '@buildwithsygma/sygma-sdk-core';
 import type { HTMLTemplateResult } from 'lit';
 import { html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import '../../../context/wallet';
 import { choose } from 'lit/directives/choose.js';
-import { when } from 'lit/directives/when.js';
 import type { Eip1193Provider } from 'packages/widget/src/interfaces';
 import type { PropertyValues } from '@lit/reactive-element';
 import {
@@ -22,7 +21,6 @@ import '../../network-selector';
 import { BaseComponent } from '../../common';
 import { Directions } from '../../network-selector/network-selector';
 import { WalletController } from '../../../controllers';
-import { tokenBalanceToNumber } from '../../../utils/token';
 import { styles } from './styles';
 
 @customElement('sygma-fungible-transfer')
@@ -103,28 +101,6 @@ export class FungibleTokenTransfer extends BaseComponent {
     }
   };
 
-  renderAmountOnDestination(): HTMLTemplateResult | null {
-    if (
-      this.transferController.selectedResource &&
-      this.transferController.pendingTransferTransaction !== undefined
-    ) {
-      const { decimals, symbol } = this.transferController.selectedResource;
-      return html`
-        <div class="amountOnDestination">
-          <span> Amount to receive: </span>
-          <span>
-            ${tokenBalanceToNumber(
-              this.transferController.resourceAmount,
-              decimals!
-            )}
-            ${symbol}
-          </span>
-        </div>
-      `;
-    }
-    return null;
-  }
-
   renderTransferStatus(): HTMLTemplateResult {
     return html` <section>
       <sygma-transfer-status
@@ -155,11 +131,12 @@ export class FungibleTokenTransfer extends BaseComponent {
             if (network) {
               this.onSourceNetworkSelected?.(network);
               this.transferController.onSourceNetworkSelected(network);
-              void this.walletController.switchEvmChain(
-                network?.chainId,
-                this.transferController.walletContext.value?.evmWallet
-                  ?.provider as Eip1193Provider
-              );
+              network.type === Network.EVM &&
+                void this.walletController.switchEvmChain(
+                  network?.chainId,
+                  this.transferController.walletContext.value?.evmWallet
+                    ?.provider as Eip1193Provider
+                );
             }
           }}
           .networks=${this.transferController.supportedSourceNetworks}
@@ -180,7 +157,8 @@ export class FungibleTokenTransfer extends BaseComponent {
         <sygma-resource-amount-selector
           .sourceDomainConfig=${this.transferController.sourceDomainConfig}
           .disabled=${!this.transferController.sourceNetwork ||
-          !this.transferController.destinationNetwork}
+          !this.transferController.destinationNetwork ||
+          this.transferController.isBuildingTransactions}
           .resources=${this.transferController.supportedResources}
           .onResourceSelected=${this.transferController.onResourceSelected}
         >
@@ -195,10 +173,8 @@ export class FungibleTokenTransfer extends BaseComponent {
         </sygma-address-input>
       </section>
       <section>
-        ${when(this.transferController.destinationAddress, () =>
-          this.renderAmountOnDestination()
-        )}
         <sygma-fungible-transfer-detail
+          .amountToReceive=${this.transferController.resourceAmountToDisplay}
           .estimatedGasFee=${this.transferController.estimatedGas}
           .selectedResource=${this.transferController.selectedResource}
           .fee=${this.transferController.fee}