+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @buildwithsygma/sygmaprotocol-widget bumped to 0.2.0
+
 ## [0.1.2](https://github.com/sygmaprotocol/sygma-widget/compare/sygmaprotocol-react-widget-v0.1.1...sygmaprotocol-react-widget-v0.1.2) (2024-04-22)